refactor(palette): extract pure run factories from useActionRunner
Move the top-level pure helpers (errorResult, makeErrorRun/StreamErrorRun/
StreamSuccessRun, jobLabel, statusFallbackAction, reduceStreamEvent + their
types) into a sibling useActionRunner/runFactories.ts. useActionRunner.ts drops
557 -> 434 non-blank/comment lines (under the 500 monolith cap). Pure refactor,
no behavior change; reduceStreamEvent re-exported for import compatibility.
